[Refactor] move ImageRenderer folder
