+++ src/pages/graduates/components/ImageRenderer.tsx
